deployments/poc-ci-pipeline-webhook-service/requirements.txt
# AI Failure Analyzer Service - Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Git Operations
GitPython==3.1.41

# LLM Integration
openai==1.10.0

# HTTP Requests
requests==2.31.0

# Environment Configuration
python-dotenv==1.0.0

# Type Hints (Python 3.11)
typing-extensions==4.9.0
